requirements.txt
flask
pyTelegramBotAPI
openai
python-binance
